Removed old vendoree branding.
