+++ payara/jacc-per-app/pom.xml
@@ -5,14 +5,14 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>org.vendoree.payara</groupId>
-        <artifactId>payara-parent</artifactId>
+        <groupId>fish.payara.samples</groupId>
+        <artifactId>samples-parent</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
 
     <artifactId>jacc-per-app</artifactId>
     <packaging>war</packaging>
-    <name>Vendor EE: Payara - Jacc Providers per Application</name>
+    <name>Payara Samples - Payara - Jacc Providers per Application</name>
 
     <dependencies>
         <dependency>
