+++ payara/jaxrs-rolesallowed/pom.xml
@@ -5,13 +5,13 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>org.vendoree.payara</groupId>
-        <artifactId>payara-parent</artifactId>
+        <groupId>fish.payara.samples</groupId>
+        <artifactId>samples-parent</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
 
     <artifactId>jaxrs-rolesallowed-EE</artifactId>
     <packaging>war</packaging>
-    <name>Vendor EE: Payara - JAX-RS using RolesAllowed and EE Security</name>
+    <name>Payara Samples - Payara - JAX-RS using RolesAllowed and EE Security</name>
 
 </project>
