+++ payara/oauth2/pom.xml
@@ -41,13 +41,13 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"> <modelVersion>4.0.0</modelVersion>
     
     <parent>
-        <groupId>org.vendoree.payara</groupId>
-        <artifactId>payara-parent</artifactId>
+        <groupId>fish.payara.samples</groupId>
+        <artifactId>samples-parent</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
     
     <artifactId>oauth2</artifactId>
-    <name>Vendor EE: Payara - OAuth2</name>
+    <name>Payara Samples - Payara - OAuth2</name>
     <packaging>war</packaging>
     
 </project>
