+++ payara/rolesPermitted/pom.xml
@@ -2,13 +2,13 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"> <modelVersion>4.0.0</modelVersion>
    
     <parent>
-        <groupId>org.vendoree.payara</groupId>
-        <artifactId>payara-parent</artifactId>
+        <groupId>fish.payara.samples</groupId>
+        <artifactId>samples-parent</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
 
     <artifactId>rolesPermitted</artifactId>
     <packaging>war</packaging>
-    <name>Vendor EE: Payara - Roles Permitted</name>
+    <name>Payara Samples - Payara - Roles Permitted</name>
     
 </project>
