+++ pom.xml
@@ -6,16 +6,12 @@
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
-	<prerequisites>
-		<maven>3.5</maven>
-	</prerequisites>
-
-	<groupId>org.vendoree</groupId>
-	<artifactId>samples-parent</artifactId>
+	<groupId>fish.payara.samples</groupId>
+	<artifactId>parent-aggregator</artifactId>
 	<version>1.0-SNAPSHOT</version>
 	<packaging>pom</packaging>
 
-	<name>EE Vendor specific samples : root</name>
+	<name>Payara Samples - root</name>
 
 	<properties>
 		<java.min.version>1.8</java.min.version>
@@ -32,7 +28,7 @@
 
 		<!-- Micro implementation versions (these are downloaded and installed
 			in these versions by Maven for the CI profiles) -->
-		<payara.version>5.183-SNAPSHOT</payara.version>
+		<payara.version>5.184</payara.version>
 		<payara.micro.version>5.182</payara.micro.version>
 		<liberty.version>17.0.0.4</liberty.version>
 		<wildfly.version>2018.4.1</wildfly.version>
