+++ test-utils/src/main/java/fish/payara/samples/Parameter.java
@@ -1,0 +1,12 @@
+package fish.payara.samples;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.FIELD})
+public @interface Parameter {
+}
+
