--- test-utils/src/main/java/org/vendoree/Parameter.java
@@ -1,12 +1,0 @@
-package org.vendoree;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.FIELD})
-public @interface Parameter {
-}
-
